Added logic for a server side events table, and system to replace more complex polling for updated data, with a simplified polling of an event table, in the SSE loop, with requests for updated data retrieving the modified games, texts, notifications and search matches based on the information in the events.
Added logic to handle updating the entire tree when a game is closed to retrieve new permissions. In the data, but also in the view--so the logic can be used at different inflexion points, where permissions change: joining a game, and closing a game.

Fixed the form endpoint for publishing, to handle store and update depending on whether the text has been saved yet.

Fixed a minor bug where game info was not updating to show unread texts from polling...